Switched the angle adjustment of the shader to ignore the part of the ST vector orthogonal to the opposite texture vector (e.g. u cross (ST - the part of ST along v))
Might need some further consideration if that's fully accurate, but it looks right? Figure out how we'd actually do the geometry to verify that in the theoretical, but for example this eliminated the gap between the connected angled and flat surfaces.

Implemented a loop-based version of the shader, that performs the minimum number of samples to catch every pixel. As expected, it's a lot worse performing (Can't really expect more than 20 lights), but it's really nice that we have this option. Upon further inspection, it's also prone to frame freezing (likely a feature of using the loops)

Also implemented a direct sample method, whereby the program repeats the full intersection calculation for several different light points. This works, but also with noticeable performance drawback.

In conclusion: For now, just running without penumbrae might be best, and for an optimized experience, some combination of the shaders is probably the ideal situation. Suppose each individual caster-light combination is split up by how big we expect the penumbra to be, and then only uses the shader necessary for it.

Next step: Cleanup (removing a lot of unused shader code, switching light radius to a parameter passed to the shader instead of one haphazardly defined there), and then implementing angle and reflection components of depth map.