Split light and shadows into separate buffers to prepare for the final render pipeline, refactored the 9 sample direct method to keep the sampling directly in the main method, but forgot to profile to see if that actually improved anything substantially. Definitely at least seemed to help a little.
Next is to split shadows into their different channels for multiple lights, the final shader to combine lights with actual color textures, and then the rendering pipeline itself is mostly done! After that comes just playing around with how we want to edit scenes so we can get a quick demo together.