Readded depth modification on a whim;
For things like bricks, where there's only one pixel set back, even with 9 samples it looks a little rapid. Could this be a case for bringing back the 4 sample smooth method, and possibly extending that to 9 samples?

Edit: After some poking around, the behavior seems inconsistent, and
should require some further analysis. For example, the brick
self-shadows seem like they might be in the right place at first, but
don't seem to change distance based on the light's distance from the
surface or how far back the etches in the bricks go, just as some kind of binary function of the two distances.